# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk Sistem Manajemen Tugas Akhir dan Skripsi Fakultas Ilmu Komputer Universitas Djuanda.",
        "title": "SIMTAS FILKOM API",
        "termsOfService": "https://filkom.unida.ac.id",
        "contact": {
            "name": "Admin SIMTAS FILKOM",
            "email": "admin@filkom.unida.ac.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/defenses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil daftar sidang (scope per role)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thesis Defense"
                ],
                "summary": "Daftar sidang",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter penguji",
                        "name": "examiner_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter tanggal mulai (RFC3339)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter tanggal akhir (RFC3339)",
                        "name": "date_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Halaman (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Per halaman (default 20)",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar sidang",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/defenses/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil detail sidang lengkap",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thesis Defense"
                ],
                "summary": "Detail sidang",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Defense ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail sidang",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.DefenseDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/defenses/{id}/result": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil hasil sidang lengkap (nilai akhir + grade + breakdown per penguji)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thesis Defense"
                ],
                "summary": "Hasil sidang",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Defense ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hasil sidang",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.DefenseResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/defenses/{id}/revision": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mencatat catatan revisi pasca sidang (Admin + Kaprodi only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thesis Defense"
                ],
                "summary": "Catat revisi sidang",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Defense ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Catatan revisi",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.revisionNotesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catatan tersimpan",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.DefenseDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Sidang belum selesai dinilai",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/defenses/{id}/schedule": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menjadwalkan/menjadwal-ulang sidang dan menunjuk penguji (Admin + Kaprodi only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thesis Defense"
                ],
                "summary": "Jadwalkan sidang",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Defense ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Jadwal + penguji",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.ScheduleDefenseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sidang terjadwal",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.DefenseDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Jadwal tidak valid",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Konflik ruangan/penguji",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/defenses/{id}/scores": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menginput nilai sidang per komponen (Dosen Penguji ditugaskan only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thesis Defense"
                ],
                "summary": "Input nilai sidang",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Defense ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nilai per komponen",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.SubmitDefenseScoreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nilai tersimpan / hasil final",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.DefenseResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Bukan penguji ditugaskan",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Sudah pernah input nilai",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/schedules/upcoming": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil seminar dan sidang terjadwal dalam N hari ke depan (Admin + Kaprodi only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thesis Defense"
                ],
                "summary": "Jadwal mendatang",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Jumlah hari ke depan (default 14)",
                        "name": "days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jadwal seminar + sidang",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/theses/{thesis_id}/defenses": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengajukan sidang skripsi (gate: seminar lulus + dokumen sidang approved) (Mahasiswa pemilik only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thesis Defense"
                ],
                "summary": "Ajukan sidang skripsi",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thesis ID (UUID)",
                        "name": "thesis_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Pengajuan dibuat",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.DefenseDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Sudah ada sidang aktif",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "422": {
                        "description": "Gate belum terpenuhi",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/theses/{thesis_id}/graduation": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menetapkan thesis berstatus graduated (Kaprodi only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Thesis Defense"
                ],
                "summary": "Tetapkan yudisium",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thesis ID (UUID)",
                        "name": "thesis_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Catatan yudisium (opsional)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/usecase.GraduationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thesis berhasil dinyatakan lulus",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "422": {
                        "description": "Prasyarat yudisium belum terpenuhi",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.revisionNotesRequest": {
            "type": "object",
            "properties": {
                "revision_notes": {
                    "type": "string"
                }
            }
        },
...
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "usecase.DefenseDetail": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "examiners": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.UserBrief"
                    }
                },
                "final_score": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "revision_notes": {
                    "type": "string"
                },
                "room": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "scores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.DefenseScoreDetail"
                    }
                },
                "status": {
                    "type": "string"
                },
                "thesis": {
                    "$ref": "#/definitions/usecase.SeminarThesisInfo"
                }
            }
        },
        "usecase.DefenseResult": {
            "type": "object",
            "properties": {
                "defense_id": {
                    "type": "string"
                },
                "examiner_scores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.ExaminerScoreBreakdown"
                    }
                },
                "final_score": {
                    "type": "number"
                },
                "grade_category": {
                    "type": "string"
                },
                "grading_components": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.GradingComponentView"
                    }
                },
                "is_complete": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "usecase.GraduationRequest": {
            "type": "object",
            "properties": {
                "notes": {
                    "type": "string"
                }
            }
        },
...
        "usecase.ScheduleDefenseRequest": {
            "type": "object",
            "required": [
                "examiner_ids",
                "room",
                "scheduled_at"
            ],
            "properties": {
                "examiner_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "room": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                }
            }
        },
...
        "usecase.SubmitDefenseScoreRequest": {
            "type": "object",
            "required": [
                "scores"
            ],
            "properties": {
                "scores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.ComponentScoreInput"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}